+++ Sources/SuperwallKit/StoreKit/StoreKitManager.swift
@@ -186,7 +186,7 @@ extension StoreKitManager {
 
   @MainActor
   private func transactionWasRestored(paywallViewController: PaywallViewController) {
-    let paywallInfo = paywallViewController.paywallInfo
+    let paywallInfo = paywallViewController.info
     Task.detached(priority: .utility) {
       let trackedEvent = InternalSuperwallEvent.Transaction(
         state: .restore,